+++ src/Plutip/Config.hs
@@ -10,7 +10,7 @@ import Cardano.Ledger.Slot (EpochSize (EpochSize, unEpochSize))
 import Data.Default (Default, def)
 import Data.Time (NominalDiffTime)
 import GHC.Generics (Generic)
-import Plutip.Launch.Extra.Types (ExtraConfig (ExtraConfig, ecEpochSize, ecSlotLength))
+import Plutip.Launch.Extra.Types (ExtraConfig (ExtraConfig, ecEpochSize, ecMaxTxSize, ecRaiseExUnitsToMax, ecSlotLength))
 
 -- | Configuration for the cluster working directory
 -- This determines where the node database, chain-index database,