+++ src/Plutip/Launch/Extra/Types.hs
@@ -1,10 +1,22 @@
 module Plutip.Launch.Extra.Types (
   ExtraConfig (..),
+  ExBudget (..),
+  stdBlockExUnits,
+  stdTxExUnits,
+  stdTxSize,
+  maxExUnits,
+  stdCollateral,
+  calculateCollateral,
 ) where
 
 import Cardano.Ledger.Slot (EpochSize)
+import Data.Aeson (FromJSON, ToJSON)
 import Data.Default (Default (def))
+import Data.Ratio ((%))
 import Data.Time (NominalDiffTime)
+import GHC.Generics (Generic)
+import Numeric.Natural (Natural)
+import PlutusCore.Evaluation.Machine.ExMemory (ExCPU (ExCPU), ExMemory (ExMemory))
 
 -- | Extra configuration options to set slot length and epoch size for local network.
 --   `ExtraConfig` used both in `PlutipConfig` and `LocalClusterConfig` to pass
@@ -16,8 +28,40 @@ import Data.Time (NominalDiffTime)
 data ExtraConfig = ExtraConfig
   { ecSlotLength :: NominalDiffTime
   , ecEpochSize :: EpochSize
+  , ecMaxTxSize :: Natural
+  , ecRaiseExUnitsToMax :: Bool
   }
   deriving stock (Show)
 
+data ExBudget = ExBudget
+  { exUnitsMem :: ExMemory
+  , exUnitsSteps :: ExCPU
+  }
+  deriving stock (Show, Eq, Generic)
+  deriving anyclass (FromJSON, ToJSON)
+
+-- below from https://github.com/input-output-hk/cardano-node/blob/master/configuration/cardano/mainnet-alonzo-genesis.json
+stdTxExUnits :: ExBudget
+stdTxExUnits = ExBudget (ExMemory 10000000) (ExCPU 10000000000)
+
+stdBlockExUnits :: ExBudget
+stdBlockExUnits = ExBudget (ExMemory 50000000) (ExCPU 40000000000)
+
+maxExUnits :: ExBudget
+maxExUnits = ExBudget (ExMemory maxBound) (ExCPU maxBound)
+
+stdCollateral :: Natural
+stdCollateral = 150
+
+stdTxSize :: Natural
+stdTxSize = 16384
+
+-- | Necessary when increasing TxSize so as not raise collateral above expected.
+calculateCollateral :: Natural -> Natural
+calculateCollateral maxTxSize =
+  if maxTxSize > stdTxSize
+    then truncate $ stdCollateral * stdTxSize % maxTxSize
+    else stdCollateral
+
 instance Default ExtraConfig where
-  def = ExtraConfig 0.2 160
+  def = ExtraConfig 0.1 80 stdTxSize False